requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
pydantic==2.5.0
python-dotenv==1.0.0
email-validator==2.0.0
requests==2.31.0